Exposed API to access other layers in InceptionV3
The tutorial code only allows saving data from the bottleneck layer which is
the last layer before classification. If I need to save data from other layers,
I need to get the tensors myself. The API allows data to be saved given the
name of the layer, which can be readily found through Tensorboard. If the layer
output is more than 1d, the output is first flattened.