Added a dropout layer to prevent overfitting
Previously overfitting was prominent when no regularization was applied. Here the traditional dropout
technique is applied and the divergence between training loss and validation loss is reduced substantially.

The paper on drop out says this technique allows an essemble of possible permutations of the same network
to be run with one graph, as the random drop on weights is a kind of regularization. Anyway the result looks
better than before.